Improved Interfaces and Node objects (#130)
Add ability for alternative primary keys to be defined for an object. This allows objects to reference other objects with a primary key that is not the actual main key over the API. E.g. You can reference an Interface by its system_id and id or name.

Refactor all node based objects to extend from a Node and Nodes classes to easy development and allow a user to reference a generic node.

Add ability for a generic Node to be converted into its actual object type: Machine, Device, RackController, or RegionController.

Add ability for a Node including all subclasses to reference the owner as an object of User.